Government_Schemes_Extraction_APIs/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
dnspython
python-dotenv
